backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
qdrant-client==1.13.3
smolagents==1.12.0
litellm==1.63.14
httpx==0.28.1
langchain-qdrant==0.2.0
langchain-ollama==0.3.0
fastembed-gpu==0.6.0
langchain_community==0.3.20
pymongo==4.11.3
langchain_openai==0.3.11
langchain-cohere==0.4.4
ragas==0.2.14
rapidfuzz==3.13.0
cohere==5.15.0
pydantic-ai==0.1.3

arize-phoenix==8.24.2
arize-phoenix-otel==0.9.1
openinference-instrumentation-langchain==0.1.40
openinference-instrumentation-smolagents==0.1.9

lightrag-hku==1.3.6

# torch==2.8.0
# transformers==4.57.0
# sentencepiece==0.2.1
FlagEmbedding==1.3.5
